+++ jablib/src/test/java/org/jabref/model/entry/MonthTest.java
@@ -6,6 +6,7 @@ import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
@@ -360,5 +361,24 @@ class MonthTest {
                 arguments("11", Month.NOVEMBER),
                 arguments("12", Month.DECEMBER)
         );
+    }
+
+    @ParameterizedTest
+    @CsvSource({
+            "6, true",
+            "12, true",
+            "#jun#, true",
+            "#jul#, true",
+            "June, false",
+            "July, false",
+            "jun, false",
+            "NotAMonth, false",
+            ", false"
+    })
+    void isStrictFormatIdentifyCorrectFormats(String input, boolean expected) {
+        if (input == null) {
+            input = "";
+        }
+        assertEquals(expected, Month.isStrictFormat(input));
     }
 }